requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-WTF==1.0.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
install==1.3.5
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-whois==0.7.3
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
